refactor(cli): extract style catalog row
Move the shared style catalog row type out of command dispatch so the
style browser no longer imports through commands.
